[update]: исправил конфигы и прочее [add]: app.manifest добавил, терь запуст идет от АДмина [remove]:
git-svn-id: https://subversion.assembla.com/svn/j_develop_station.3/trunk@33 ed5da3e8-8e1b-49b9-9390-93002e306179

+++ client/classes/config/RConfig.cs
@@ -74,7 +74,8 @@ namespace com.jds.AWLauncher.classes.config
             try
             {
                 RegistryKey key = Registry.CurrentUser;
-                key.DeleteSubKeyTree("Software\\J Develop Station\\GUpdater");
+                if (key.OpenSubKey("Software\\J Develop Station\\GUpdater") != null)
+                    key.DeleteSubKeyTree("Software\\J Develop Station\\GUpdater");
             }
             catch
             {
